requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
imageio-ffmpeg==0.4.9
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
libretranslatepy==2.1.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
multiprocess==0.70.16
mysql-connector-python==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.2
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py-cord==2.5.0
pycord==0.1.1
pycparser==2.22
PyExecJS==1.5.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
translate==3.6.1
translators==5.9.2
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.9.4
